requirements.txt
numpy
boto3
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
pandas
requests